requirements.txt
Django==5.0.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.2.0
openpyxl==3.1.2
pytz==2024.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
